split route-bounds.ts by concept
route-bounds.ts had grown to cover three loosely-related concerns:
GeoJSON conversion / per-step geometry, bbox-to-corner-pair helpers
for camera fitting, and route preview/summary display helpers. It also
held a roundTrailerPoint helper that wasn't actually a route concept
and had no production callers.

Split into:
- route-features.ts -- toRouteFeatures, bearingAfterStepManeuver
  (callers: controllers/app.ts, create-app-handlers.ts).
- route-bounds.ts -- bboxToCornerPair, bboxesToCornerPairs,
  routeCornerPair, routesCornerPairs (the actual bounds concept;
  callers: create-app-handlers.ts, create-app-reactions.ts).
- route-display.ts -- sortedRoutePreviewIndices, toRouteSummary
  (callers: create-app-reactions.ts; toRouteSummary will be wired
  into create-app.tsx in the next commit).

Drop roundTrailerPoint -- its inline equivalent in create-app.tsx
already does the rounding, and the named helper had drifted into
orphan status.

Tests split correspondingly into route-features.test.ts,
route-display.test.ts, and a slimmed route-bounds.test.ts.
